+++ tests/010-sup-start_stop.cpp
@@ -79,7 +79,7 @@ struct sample_plugin_t : r::plugin::plugin_base_t {
 
     static const void *class_identity;
 
-    const void *identity() const noexcept override { return  class_identity; }
+    const void *identity() const noexcept override { return class_identity; }
 
     void activate(r::actor_base_t *actor_) noexcept override {
         parent_t::activate(actor_);
@@ -88,15 +88,12 @@ struct sample_plugin_t : r::plugin::plugin_base_t {
 
     void deactivate() noexcept override { parent_t::deactivate(); }
 
-    void on_message(message::sample_payload_t& ) noexcept {
-        message_received = true;
-    }
+    void on_message(message::sample_payload_t &) noexcept { message_received = true; }
 
     bool message_received = false;
 };
 
 const void *sample_plugin_t::class_identity = &sample_plugin_t::class_identity;
-
 
 struct sample_sup2_t : public rt::supervisor_test_t {
     using sup_base_t = rt::supervisor_test_t;
@@ -255,9 +252,19 @@ struct sample_actor5_t : public rt::actor_test_t {
         send<payload::sample_payload_t>(get_address());
         send<payload::sample_payload_t>(get_address());
     }
-
 };
 
+struct sample_actor6_t : public rt::actor_test_t {
+    using rt::actor_test_t::actor_test_t;
+
+    void on_start() noexcept override {
+        rt::actor_test_t::on_start();
+        start_timer(r::pt::minutes(1), *this, &sample_actor6_t::on_timer);
+    }
+
+    void on_timer(r::request_id_t, bool cancelled) noexcept { this->cancelled = cancelled; }
+    bool cancelled = false;
+};
 
 TEST_CASE("on_initialize, on_start, simple on_shutdown (handled by plugin)", "[supervisor]") {
     destroyed = 0;
@@ -449,7 +456,7 @@ TEST_CASE("io tagging (in plugin) & intercepting", "[actor]") {
     CHECK(sup->counter == 2);
     auto plugin = act->access<rt::to::get_plugin>(sample_plugin_t::class_identity);
     CHECK(plugin);
-    CHECK(static_cast<sample_plugin_t*>(plugin)->message_received);
+    CHECK(static_cast<sample_plugin_t *>(plugin)->message_received);
 
     sup->do_shutdown();
     sup->do_process();
@@ -458,3 +465,19 @@ TEST_CASE("io tagging (in plugin) & intercepting", "[actor]") {
     CHECK(sup->get_state() == r::state_t::SHUT_DOWN);
 }
 
+TEST_CASE("timers cancellation", "[actor]") {
+    r::system_context_ptr_t system_context = new r::system_context_t();
+    auto sup = system_context->create_supervisor<rt::supervisor_test_t>().timeout(rt::default_timeout).finish();
+    auto act = sup->create_actor<sample_actor6_t>().timeout(rt::default_timeout).finish();
+    sup->do_process();
+    CHECK(act->get_state() == r::state_t::OPERATIONAL);
+    CHECK(sup->get_state() == r::state_t::OPERATIONAL);
+    CHECK(!act->access<rt::to::timers_map>().empty());
+
+    sup->do_shutdown();
+    sup->do_process();
+
+    CHECK(act->get_state() == r::state_t::SHUT_DOWN);
+    CHECK(sup->get_state() == r::state_t::SHUT_DOWN);
+    CHECK(act->access<rt::to::timers_map>().empty());
+}