+++ tests/142-thread_timer.cpp
@@ -47,6 +47,7 @@ struct bad_actor_t : public r::actor_base_t {
         auto sup = static_cast<rth::supervisor_thread_t *>(supervisor);
         sup->updat_time();
         start_timer(r::pt::milliseconds(1), *this, &bad_actor_t::delayed_start);
+        start_timer(r::pt::minutes(1), *this, &bad_actor_t::delayed_start); // to be cancelled
     }
 
     void delayed_start(r::request_id_t, bool) noexcept {